+++ jkorm-spring/src/test/java/com/tang/jkorm/spring/mapper/JavaAccountMapper.java
@@ -1,0 +1,15 @@
+package com.tang.jkorm.spring.mapper;
+
+import com.tang.jkorm.mapper.BaseMapper;
+import com.tang.jkorm.spring.entity.Account;
+
+/**
+ * @author Tang
+ */
+public interface JavaAccountMapper extends BaseMapper<Account> {
+
+    default int insertAccount(Account account) {
+        return insert(account);
+    }
+
+}