+++ jkorm-spring/src/test/kotlin/com/tang/jkorm/spring/Main.kt
@@ -1,5 +1,0 @@
-package com.tang.jkorm.spring
-
-/**
- * @author Tang
- */