+++ jkorm-spring/src/test/kotlin/com/tang/jkorm/spring/entity/Account.kt
@@ -1,0 +1,19 @@
+package com.tang.jkorm.spring.entity
+
+import com.tang.jkorm.annotation.Id
+
+/**
+ * @author Tang
+ */
+class Account (
+
+    @Id
+    var id: Long? = null,
+    var username: String = "",
+    var password: String = ""
+
+) {
+    override fun toString(): String {
+        return "'Account => id=$id username=$username password=$password'"
+    }
+}