+++ jkorm-spring/src/test/kotlin/com/tang/jkorm/spring/mapper/AccountMapper.kt
@@ -1,0 +1,15 @@
+package com.tang.jkorm.spring.mapper
+
+import com.tang.jkorm.mapper.BaseMapper
+import com.tang.jkorm.spring.entity.Account
+
+/**
+ * @author Tang
+ */
+interface AccountMapper : BaseMapper<Account> {
+
+    fun insertAccount(account: Account): Int {
+        return insert(account)
+    }
+
+}